pyproject.toml
[project]
name = "ssi-backend"
version = "0.1.0"
description = "The Backend of the Service Status Indicator monitoring framework"
authors = [
    {name = "RemiZlatinis",email = "remizlatinis@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.6,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "channels[daphne] (>=4.3.1,<5.0.0)",
    "django-allauth (>=65.11.2,<66.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "dj-rest-auth (>=7.0.1,<8.0.0)",
    "django-ratelimit (>=4.0.0,<5.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "channels-redis (>=4.3.0,<5.0.0)",
    "dj-database-url (>=3.0.1,<4.0.0)",
    "servestatic (>=3.1.0,<4.0.0)",
    "requests (>=2.32.5,<3.0.0)",
]

[tool.poetry]
package-mode = false
packages = [{include = "project"}]

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
ruff = "^0.13.0"
django-debug-toolbar = "^6.0.0"
pre-commit = "^3.8.0"
mypy = "^1.18.2"
django-stubs = {extras = ["compatible-mypy"], version = "^5.2.5"}
djangorestframework-stubs = "^3.16.3"

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.ruff]
# Set the same line length as Black to ensure they agree.
line-length = 88

[tool.ruff.lint]
# Enable a good set of default rules, plus Django-specific ones.
select = [
    "E",  # pycodestyle errors
    "F",  # Pyflakes
    "W",  # pycodestyle warnings
    "I",  # isort (import sorting)
    "DJ", # django-specific rules
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
